Get very basic drawing application working
Set up dev environment, not ready for running this stuff on travis yet
so the build will probably (almost certainly) fail.

Investigated webpack hot module reloading, but it doesn't play nicely
with nodemon and seems more effort than its worth for this project so
I've ditched it in favour of live reload which seems to work quite
nicely.

Currently there is a mix of `require()` (CommonJS) and `import X from
'X'` style (ES7) imports which should probably be normalised to the ES7
style, but I haven't set up webpack for the server side stuff yet