Add deploy configuration to server and package.json
The server conditionally checks whether it's running in development or
not, if in dev we use webpack middle ware to compile the js files. If
we're in production, we should only be using the built versions (built
using the command `webpack`, which is issued on `npm run postinstall`).

By adding `webpack` to the `postinstall` target in `package.json` Heroku
should run webpack in it's slug creation process prior to deployment